Only define `cohort(x)` for finite numbers
While we also defined the cohort of +inf/-inf/NaN, those values were
never passed to `cohort()` and thus all the algorithms (correctly)
assume that the result of `cohort` is either a zero or a finite
mathematical value. While this is true, it is difficult to read them
because the type definition of  `cohort` says that it can also return
`+inf/-inf/NaN`. This updates the type definition to match usage,
rather than unnecessarily accepting more inputs.

There is one case in .add and friends where we are actually passing
+inf/-inf/NaN to cohort(), but that is just a bug since we then do not
handle the case in which it returns +inf/-inf/NaN.