Fix most known bugs, including separation of selection of quantum from rounding, as well as API additions (#159)
* Get started using the suggested `PickQuantum` AO

As suggested in #122

* Fix and simplify .add, .subtract and .multiply

There were a bunch of errors with handling of zeroes
and infinities. This commit fixes those, and reduces
the number of branches to make the algorithm shorter
and simpler to follow.

* Only define `cohort(x)` for finite numbers

While we also defined the cohort of +inf/-inf/NaN, those values were
never passed to `cohort()` and thus all the algorithms (correctly)
assume that the result of `cohort` is either a zero or a finite
mathematical value. While this is true, it is difficult to read them
because the type definition of  `cohort` says that it can also return
`+inf/-inf/NaN`. This updates the type definition to match usage,
rather than unnecessarily accepting more inputs.

There is one case in .add and friends where we are actually passing
+inf/-inf/NaN to cohort(), but that is just a bug since we then do not
handle the case in which it returns +inf/-inf/NaN.

* Define rounding and use it, along with pick-quantum, throughout

* Ensure .abs returns a non-negative value

* Fix `isFinite`

* Remove references to "rational number"

(Of course, an alternative would be to define the term.)

* Handle second argument NaN in .compare

* round: Do `RequireInternalSlot` before validating arguments

* Unconditionally throw in `valueOf`

Don't even ensure that we have the right internal slot.

* Set variable rather than return early

We need to ensure that `new Number(...)` returns a new
object, rather than a primitive.

* Ensure variable is defined

* Use the term "real number"

* Use `abs(...)` rather than `|...|`

* Use assertion rather than note

* Add <sub>𝔻</sub> after pairs

* Fix return type of ToIntlMathematicalValue

We don't return a decimal.

* Handle returning 0 from remainder

* Qualify argument of rounding (shouldn't be zero)

* Use proper IDs

* When rounding, enumerate 34 significant digits and consult the 35th

* Show where we overflow to infinity in rounding

* Handle denormalized values in rounding

* Add IDs for normalized/denormalized

* Handle underflow with increasingly coarse-grained approximations

* Handle the possibility that rounding returns 0

* Fix busted <sub>D</sub> and calls to `PickQuantum`

* In scale10, allow NaN and infinity as argument

Throw only if we are given a non-integer finite argument.

* Return a +/-0 from RoundToDecimal128Domain

* Fix quantum of division

* Define a separate AO for rounding positive numbers

* Embed utility rounding AO inside the other one

Also, fix a positive infinity sneaking out.

---------

Co-authored-by: Nicolò Ribaudo <nribaudo@igalia.com>
Co-authored-by: Nicolò Ribaudo <hello@nicr.dev>